refactor(app): extract search & replace orchestrators to search_ops (phase 3a)
Move the 1307-line Search and Replace method cluster out of render.rs
into a new app/search_ops.rs sibling file. Contents:

  Query/overlay:   clear_search_highlights, clear_search_overlays,
                   update_search_highlights, refresh_search_overlays,
                   check_search_overlay_refresh
  Regex build:     build_search_regex (delegates to regex_replace),
                   build_replace_regex, get_regex_match_len,
                   expand_regex_replacement
  Scan entry:      start_search_scan, get_search_match_positions
  Navigation:      move_cursor_to_match, find_next, find_previous,
                   find_match_in_direction, find_selection_next,
                   find_selection_previous,
                   get_selection_or_word_for_search_with_pos
  Search flow:     perform_search, finalize_search
  Replace:         perform_replace
  Interactive replace: start_interactive_replace,
                   handle_interactive_replace_key,
                   find_next_match_for_replace, replace_current_match,
                   move_to_current_match, finish_interactive_replace

Also removes the `SearchDirection` enum from render.rs (it was only used
by this cluster) and the now-unused `AnyhowResult` import.

No behavioural change; this is Strategy A from the plan (sibling
impl Editor file) applied to the largest cohesive cluster in render.rs.
render.rs goes from 5207 -> 3899 lines (-1308).
